refactor: merit button + hook
